stantion_back/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.2.0
psycopg2==2.9.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
